rewrote the regexp layer in pure C to avoid binary incompatibility issues with ICU
git-svn-id: http://svn.macosforge.org/repository/ruby/MacRuby/trunk@4434 23306eb0-4c56-4727-a40e-e92c0eb68959

+++ compiler.cpp
@@ -4802,7 +4802,7 @@ RoxorAOTCompiler::compile_init_function(void)
 	    case T_REGEXP:
 		{
 		    const UChar *chars = NULL;
-		    long chars_len = 0;
+		    int32_t chars_len = 0;
 
 		    regexp_get_uchars(val, &chars, &chars_len);
